+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # UNRELEASED
 - ApiServer: added `--requestheader-client-ca-file` cmdline flag
 - ControllerManager: explicit HPA settings
+- New addon: [node problem detector](https://github.com/kubernetes/node-problem-detector)
+  Disabled by default, use `node['kubernetes']['addons']['npd']['enabled']` to activate
 - [Kubernetes: 1.11.2](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.11.md#changelog-since-v1111)
 - [cri-o: 1.11.2](https://github.com/kubernetes-incubator/cri-o/releases/tag/v1.11.2)
 - [buildah: 1.3](https://github.com/projectatomic/buildah/releases/tag/v1.3)